+++ packages/bigquery/index.cjs
@@ -1,4 +1,5 @@
-const {BigQuery} = require('@google-cloud/bigquery');
+const { BigQuery } = require('@google-cloud/bigquery');
+const { EvidenceType, TypeFidelity } = require('@evidence-dev/db-commons');
 
 const standardizeResult = async(result) => {
     var output = [];
@@ -46,8 +47,6 @@ const getCredentials = async(database) => {
     }
 } 
 
-
-
 const runQuery = async (queryString, database) => {
     try {
         const credentials = await getCredentials(database)
@@ -55,8 +54,11 @@ const runQuery = async (queryString, database) => {
 
         const [job] = await connection.createQueryJob({query: queryString });
         const [rows] = await job.getQueryResults();
-        const standardizedResult = await standardizeResult(rows)
-        return standardizedResult
+        const [, , response] = await job.getQueryResults({
+            autoPaginate: false,
+        });
+        const standardizedRows = await standardizeResult(rows);
+        return { rows: standardizedRows, columnTypes : mapResultsToEvidenceColumnTypes(response) };
     } catch (err) {
         if (err.errors) {
             throw err.errors[0].message           
@@ -65,5 +67,59 @@ const runQuery = async (queryString, database) => {
         }        
     }
 };
+
+const nativeTypeToEvidenceType = function (nativeFieldType, defaultType = undefined) {
+    switch (nativeFieldType) {
+        case 'BOOL':
+        case 'BOOLEAN':
+          return EvidenceType.BOOLEAN;
+        case 'INT64':
+        case 'INT':
+        case 'SMALLINT':
+        case 'INTEGER':
+        case 'BIGINT':
+        case 'TINYINT':
+        case 'BYTEINT':
+        case 'DECIMAL':
+        case 'NUMERIC':
+        case 'BIGDECIMAL':
+        case 'BIGNUMERIC':
+        case 'FLOAT64':
+          return EvidenceType.NUMBER;
+        case 'STRING':
+          return EvidenceType.STRING;
+        case 'TIME':
+        case 'TIMESTAMP':
+        case 'DATE':
+        case 'DATETIME':
+          return EvidenceType.DATE;
+        case 'STRUCT':
+        case 'ARRAY':
+        case 'BYTES':
+        case 'GEOGRAPHY':
+        case 'INTERVAL':
+        case 'JSON':
+        default:
+          return defaultType;
+    }
+};
+
+const mapResultsToEvidenceColumnTypes = function (results) {
+    return results?.schema?.fields?.map(field => {
+        let typeFidelity = TypeFidelity.PRECISE;
+        let evidenceType = nativeTypeToEvidenceType(field.type);
+        if (!evidenceType) {
+            typeFidelity = TypeFidelity.INFERRED;
+            evidenceType = EvidenceType.STRING;
+        }
+        return (
+          {
+            'name': field.name,
+            'evidenceType': evidenceType,
+            'typeFidelity': typeFidelity,
+          });
+    });
+};
+
 
 module.exports = runQuery