+++ packages/db-orchestrator/index.cjs
@@ -4,31 +4,30 @@ const chalk = require('chalk')
 const logEvent = require('@evidence-dev/telemetry')
 const readline = require('readline');
 
-const getCache = function (dev, queryString, queryTime, settings) {
-    queryTime = md5(queryTime)
-    if (dev) {
-        const cache = readJSONSync("./.evidence-queries/cache/" + queryTime + "/" + md5(queryString) + ".json", { throws: false })
-        if (cache) {
-            logEvent("cache-query", dev, settings)
-            return cache
-        }
+const cacheDirectory = "./.evidence-queries/cache";
+
+const getQueryCachePaths = (queryString, queryTime) => {
+    let queryTimeMD5 = md5(queryTime);
+    let queryStringMD5 = md5(queryString);
+    let path = `${cacheDirectory}/${queryTimeMD5}`;
+    return {
+        'cacheDirectory': path,
+        'resultsCacheFile': `${cacheDirectory}/${queryTimeMD5}/${queryStringMD5}.json`,
+        'columnTypeCacheFile': `${cacheDirectory}/${queryTimeMD5}/${queryStringMD5}-column-types.json`,
     }
 }
 
-const updateCache = function (dev, queryString, data, queryTime) {
-    queryTime = md5(queryTime)
-    if (dev) {
-        if (!pathExistsSync("./.evidence-queries")) {
-            mkdirSync("./.evidence-queries")
+const updateCache = function (devMode, queryString, data, columnTypes, queryTime) {
+    if (devMode) {
+        const {cacheDirectory, resultsCacheFile, columnTypeCacheFile} = getQueryCachePaths(queryString, queryTime);
+        if (!pathExistsSync(cacheDirectory)) {
+            emptyDirSync(cacheDirectory);
+            mkdirSync(cacheDirectory, { recursive: true });
         }
-        if (!pathExistsSync("./.evidence-queries/cache/")) {
-            mkdirSync("./.evidence-queries/cache/")
+        writeJSONSync(resultsCacheFile, data, { throws: false });
+        if (columnTypes) {
+            writeJSONSync(columnTypeCacheFile, columnTypes, { throws: false });
         }
-        if (!pathExistsSync("./.evidence-queries/cache/" + queryTime)) {
-            emptyDirSync('./.evidence-queries/cache/')
-            mkdirSync("./.evidence-queries/cache/" + queryTime)
-        }
-        writeJSONSync("./.evidence-queries/cache/" + queryTime + "/" + md5(queryString) + ".json", data, { throws: false })
     }
 }
 
@@ -47,9 +46,6 @@ const validateQuery = function (query) {
     }
 }
 
-
-
-
 const importDBAdapter = async function(settings) {
     try {
         databaseType = settings ? settings.database : process.env["DATABASE"] || process.env["database"]
@@ -63,6 +59,14 @@ const importDBAdapter = async function(settings) {
     }
 }
 
+/** adds columnTypes to metadata in the page `data` object */
+const populateColumnTypeMetadata = (data, queryIndex, columnTypes) => {
+    let queryMetaData = data.evidencemeta?.queries?.[queryIndex];
+    if (queryMetaData && columnTypes) {
+        queryMetaData.columnTypes = columnTypes;
+    }
+} 
+
 const runQueries = async function (routeHash, dev) {
     const settings = readJSONSync('./evidence.settings.json', {throws:false})
     const runQuery = await importDBAdapter(settings)
@@ -75,20 +79,38 @@ const runQueries = async function (routeHash, dev) {
     if (queries.length > 0) {
         let data = {}
         data["evidencemeta"] = {queries} // eventually move to seperate metadata API (md frontmatter etc.) 
-        for (let query of queries) {
+        for (let queryIndex in queries) {
+            let query = queries[queryIndex];
             let queryTime = new Date(new Date().getFullYear(), new Date().getMonth(), new Date().getDate(), new Date().getHours());              
-            let cache = getCache(dev, query.compiledQueryString, queryTime, settings)
+
+            let cache, columnTypeCache;
+            if (dev) {
+                const { resultsCacheFile, columnTypeCacheFile } = getQueryCachePaths(query.compiledQueryString, queryTime);
+                cache = readJSONSync(resultsCacheFile, { throws: false });
+                columnTypeCache = readJSONSync(columnTypeCacheFile, { throws: false });
+            }
             if (cache) {
-                data[query.id] = cache
-                process.stdout.write(chalk.greenBright("✓ "+ query.id) +  chalk.grey(" from cache \n"))
+                logEvent("cache-query", dev, settings);
+                data[query.id] = cache;
+                if (columnTypeCache) {
+                    populateColumnTypeMetadata(data, queryIndex, columnTypeCache);
+                }
+                process.stdout.write(chalk.greenBright("✓ "+ query.id) +  chalk.grey(" from cache \n"));
             } else {
                 try {
-                    process.stdout.write(chalk.grey("  "+ query.id +" running..."))
-                    validateQuery(query)
-                    data[query.id] = await runQuery(query.compiledQueryString, settings?.credentials, dev)
+                    process.stdout.write(chalk.grey("  "+ query.id +" running..."));
+                    validateQuery(query);
+
+                    let {rows, columnTypes} = await runQuery(query.compiledQueryString, settings?.credentials, dev);
+
+                    data[query.id] = rows;
+                    populateColumnTypeMetadata(data, queryIndex, columnTypes);
+
                     readline.cursorTo(process.stdout, 0);
                     process.stdout.write(chalk.greenBright("✓ "+ query.id) + chalk.grey(" from database \n"))
-                    updateCache(dev, query.compiledQueryString, data[query.id], queryTime)
+
+                    updateCache(dev, query.compiledQueryString, data[query.id], columnTypes, queryTime);
+
                     logEvent("db-query", dev, settings)
                 } catch(err) {
                     readline.cursorTo(process.stdout, 0);
@@ -115,7 +137,7 @@ const testConnection = async function () {
 
     try {
         process.stdout.write(chalk.grey("  "+ query.id +" running..."))
-        queryResult = await runQuery(query.compiledQueryString, settings.credentials)
+        await runQuery(query.compiledQueryString, settings.credentials)
         readline.cursorTo(process.stdout, 0);
         process.stdout.write(chalk.greenBright("✓ "+ query.id) + chalk.grey(" from database \n"))
         result = "Database Connected";