+++ sites/example-project/src/pages/api/[route].json.js
@@ -3,7 +3,7 @@ import { dev } from '$app/env';
 
 export async function get({params}) {
   const { route } = params;
-  const data = await runQueries(route, dev)
+  const data = await runQueries(route, dev);
   return {
       body: {
         data